Move binding logic to BindsManager
