+++ engine/src/main/java/org/terasology/config/InputConfig.java
@@ -19,14 +19,9 @@ import org.terasology.context.Context;
 
 public class InputConfig {
 
-    private BindsConfig binds = new BindsConfig();
     private ControllerConfig controllers = new ControllerConfig();
     private float mouseSensitivity;
     private boolean mouseYAxisInverted;
-
-    public BindsConfig getBinds() {
-        return binds;
-    }
 
     public ControllerConfig getControllers() {
         return controllers;
@@ -41,8 +36,6 @@ public class InputConfig {
     }
 
     public void reset(Context context) {
-        binds.setBinds(BindsConfig.createDefault(context));
-
         Config defaultConfig = new Config();
         defaultConfig.loadDefaults();
         InputConfig defaultInputConfig = defaultConfig.getInput();
